moved Chat*Controllers into controller folder
